requirements.txt
# Efficient Vector Similarity Search (course project)
# HNSW via hnswlib; embeddings: sentence-transformers, torchvision, numpy/scipy
hnswlib>=0.8.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
psutil>=5.9.0
sentence-transformers>=2.2.0
torch>=2.0.0
torchvision>=0.15.0
requests>=2.31.0
faiss-cpu>=1.7.4
matplotlib>=3.8.0
